プロフィールページ作成

+++ api/app/controllers/api/v1/posts_controller.rb
@@ -1,4 +1,5 @@
 class Api::V1::PostsController < ApplicationController
+    before_action :authenticate_api_v1_user!, only: [:create, :update, :destroy]
     def index
         posts = Post.all.order(created_at: :desc)
         render json: posts
@@ -20,21 +21,29 @@ class Api::V1::PostsController < ApplicationController
 
     def update
         post = Post.find(params[:id])
-        if post.update(post_params)
-            render json: post
+        if current_api_v1_user.id == post.user_id
+            if post.update(post_params)
+                render json: post
+            else
+                render json: post.errors, status: 422
+            end
         else
-            render json: post.errors, status: 422
+            render json: {message: 'can not update data'}, status: 422
         end
     end
 
     def destroy
         post = Post.find(params[:id])
-        post.destroy
-        render json: post
+        if current_api_v1_user.id == post.user_id
+            post.destroy
+            render json: post
+        else
+            render json: {message: 'can not delete data'}, status: 422
+        end
     end
 
     private
     def post_params
-        params.require(:post).permit(:content)
+        params.require(:post).permit(:content).merge(user_id: current_api_v1_user.id)
     end
 end
